Added Merchant Ranking

+++ src/main/java/com/ecommerce/merchant/DTO/ProductMerchantDTO.java
@@ -6,14 +6,14 @@ public class ProductMerchantDTO {
     private String productId;
     private double price;
     private int quantity;
-    private int rank;
+    private double rank;
     private double rating;
 
-    public int getRank() {
+    public double getRank() {
         return rank;
     }
 
-    public void setRank(int rank) {
+    public void setRank(double rank) {
         this.rank = rank;
     }
